MAINT: change the name of staging management cluster
to follow naming convention so its the same as prod-mgmt

This is a breaking change! - don't merge this until we're ready to run ./deploy again on the staging cluster. This is part of a test we need to document to re-make an existing cluster

staging worker cluster will need to be recreated too