MAINT: Switch to manual replica count
Since these are dedicated nodes, there's no point autoscaling since
nothing will spin down.

Instead have some initial guesses for our replicas that should be safe
everywhere. Once this is deployed we can view how it fills the nodes
and override it for prod / staging, depending on the capacity of the
node