MAINT: refactor the way we store sops secrets
Split sops secrets into 2 parts - infra and apps

the infra secrets should only be used by the management cluster

new apps folder for storing cluster-specific secrets pertaining to apps running on the cluster.

This allows us to use different keys for the child clusters and the secrets will not be accessible by the management cluster - allows separation of roles.