MAINT: Edit values of dev application
Reduce replicas to 0 so the app isn't running anymore. Change app version to reflect actual app version. Change channel to dev workspace channel. Change secrets to use dev slack app rather than prod. Change image to the staging project and leave the tag as a placeholder to be changed when more work is done on dev