create self-signed certs for opensearch
use cert-manager to create self-signed certs for the different opensearch components to talk to each other.

Previously this was achieved by setting generate:true for http and transport layers.

This didn't play nicely with argocd - see this issue - https://github.com/bitnami/charts/issues/24643. The chart uses helm lookup which is not supported in argocd.
This causes argocd to regenerate these certificates during every update - (every 5m) - which has a significant slow down

This PR fixes that by generating our own certs. I placed this as part of argocd chart because it only needs to be done for argocd so I didn't want to upstream it to our charts repo